avocado-server.spec: allow the commit to be passed on the command line
This allows us to build RPM packages for any commit/version of
avocado-server.  The use case here is to be able to build release
versions (by default commit should be pointing to the latest release),
or if passed on the command line, an specific commit, such as the
latest development state.

Signed-off-by: Cleber Rosa <crosa@redhat.com>