avocado-server daemon: use avocadoserver user and groups
It's a good practice to have specific users and groups for daemons, so
that its data is made private and possible damages caused by bugs,
including security related ones, are better contained.

This commit introduces the creation of the "avocadoserver" user
and group, following the Fedora guidelines.

Also the gunicorn WSGI/HTTP server is now set to drop its initial root
privileges to the privileges of the avocadoserver user and group.

Signed-off-by: Cleber Rosa <crosa@redhat.com>